+++ libfabric.vcxproj.filters
@@ -363,6 +363,9 @@
     <ClCompile Include="prov\rxm\src\rxm_msg.c">
       <Filter>Source Files\prov\rxm\src</Filter>
     </ClCompile>
+    <ClCompile Include="prov\rxm\src\rxm_tagged.c">
+      <Filter>Source Files\prov\rxm\src</Filter>
+    </ClCompile>
     <ClCompile Include="src\iov.c">
       <Filter>Source Files\src</Filter>
     </ClCompile>
