+++ prov/rxm/Makefile.include
@@ -7,6 +7,7 @@ _rxm_files = \
        prov/rxm/src/rxm_conn.c	\
        prov/rxm/src/rxm_ep.c	\
        prov/rxm/src/rxm_msg.c	\
+       prov/rxm/src/rxm_tagged.c	\
        prov/rxm/src/rxm_cq.c	\
        prov/rxm/src/rxm_av.c	\
        prov/rxm/src/rxm_rma.c	\
